EV_app_final/requirements.txt
pandas==2.2.3
numpy==2.2.2
matplotlib==3.10.0
seaborn==0.13.2
scikit-learn==1.6.1
streamlit==1.44.1
SQLAlchemy==2.0.38
openpyxl==3.1.5
pillow==11.1.0
